feat: continued components
#38
