+++ src/components/CreateEntry.test.tsx
@@ -1,0 +1,7 @@
+import { render } from "@testing-library/react"
+import { CreateEntry } from "./CreateEntry"
+
+// FIXME #38 add tests
+test("Renders the component", () => {
+  render(<CreateEntry />)
+})
