+++ src/components/ProjectSelector.test.tsx
@@ -1,0 +1,6 @@
+import { render } from "@testing-library/react"
+import { ProjectSelector } from "./ProjectSelector"
+
+test("renders the component", () => {
+  render(<ProjectSelector />)
+})
