+++ src/components/ProjectSelector.tsx
@@ -1,0 +1,12 @@
+import { Select } from "antd"
+
+const { Option } = Select
+
+// FIXME #38 implement
+export function ProjectSelector() {
+  return (
+    <Select style={{ width: 400 }} tabIndex={2}>
+      <Option value="1">xxx</Option>
+    </Select>
+  )
+}
