+++ src/components/TaskSelector.test.tsx
@@ -1,0 +1,7 @@
+import { render } from "@testing-library/react"
+import { TaskSelector } from "./TaskSelector"
+
+// FIXME #38 implement
+test("render the component", () => {
+  render(<TaskSelector />)
+})
